backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
black==25.11.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
multidict==6.7.0
mypy-extensions==1.1.0
openai==2.8.1
packaging==25.0
pathspec==0.12.1
pip @ file:///Applications/PyCharm.app/Contents/plugins/python-ce/helpers/pip-24.3.1-py2.py3-none-any.whl
platformdirs==4.5.0
postgrest==2.24.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pygame==2.6.1
pyjwt==2.10.1
python-dotenv==1.2.1
pytokens==0.3.0
realtime==2.24.0
sniffio==1.3.1
storage3==2.24.0
strenum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
websockets==15.0.1
yarl==1.22.0
